refactor to use hooks
