+++ mlflow/server/js/src/experiment-tracking/pages/experiment-overview/components/TraceRequestsChart.tsx
@@ -1,10 +1,8 @@
-import React, { useMemo, useCallback } from 'react';
+import React from 'react';
 import { useDesignSystemTheme, ChartLineIcon } from '@databricks/design-system';
 import { FormattedMessage } from 'react-intl';
 import { BarChart, Bar, XAxis, YAxis, Tooltip, ResponsiveContainer } from 'recharts';
-import { MetricViewType, AggregationType, TraceMetricKey } from '@databricks/web-shared/model-trace-explorer';
-import { useTraceMetricsQuery } from '../hooks/useTraceMetricsQuery';
-import { formatTimestampForTraceMetrics, useTimestampValueMap } from '../utils/chartUtils';
+import { useTraceRequestsChartData } from '../hooks/useTraceRequestsChartData';
 import {
   OverviewChartLoadingState,
   OverviewChartErrorState,
@@ -17,54 +15,13 @@ import {
 } from './OverviewChartComponents';
 import type { OverviewChartProps } from '../types';
 
-export const TraceRequestsChart: React.FC<OverviewChartProps> = ({
-  experimentId,
-  startTimeMs,
-  endTimeMs,
-  timeIntervalSeconds,
-  timeBuckets,
-}) => {
+export const TraceRequestsChart: React.FC<OverviewChartProps> = (props) => {
   const { theme } = useDesignSystemTheme();
   const tooltipStyle = useChartTooltipStyle();
   const xAxisProps = useChartXAxisProps();
 
-  // Fetch trace count metrics grouped by time bucket
-  const {
-    data: traceCountData,
-    isLoading,
-    error,
-  } = useTraceMetricsQuery({
-    experimentId,
-    startTimeMs,
-    endTimeMs,
-    viewType: MetricViewType.TRACES,
-    metricName: TraceMetricKey.TRACE_COUNT,
-    aggregations: [{ aggregation_type: AggregationType.COUNT }],
-    timeIntervalSeconds,
-  });
-
-  const traceCountDataPoints = useMemo(() => traceCountData?.data_points || [], [traceCountData?.data_points]);
-
-  // Get total requests
-  const totalRequests = useMemo(
-    () => traceCountDataPoints.reduce((sum, dp) => sum + (dp.values?.[AggregationType.COUNT] || 0), 0),
-    [traceCountDataPoints],
-  );
-
-  // Create a map of counts by timestamp using shared utility
-  const countExtractor = useCallback(
-    (dp: { values?: Record<string, number> }) => dp.values?.[AggregationType.COUNT] || 0,
-    [],
-  );
-  const countByTimestamp = useTimestampValueMap(traceCountDataPoints, countExtractor);
-
-  // Prepare chart data - fill in all time buckets with 0 for missing data
-  const chartData = useMemo(() => {
-    return timeBuckets.map((timestampMs) => ({
-      name: formatTimestampForTraceMetrics(timestampMs, timeIntervalSeconds),
-      count: countByTimestamp.get(timestampMs) || 0,
-    }));
-  }, [timeBuckets, countByTimestamp, timeIntervalSeconds]);
+  // Fetch and process requests chart data
+  const { chartData, totalRequests, isLoading, error, hasData } = useTraceRequestsChartData(props);
 
   if (isLoading) {
     return <OverviewChartLoadingState />;
@@ -85,7 +42,7 @@ export const TraceRequestsChart: React.FC<OverviewChartProps> = ({
 
       {/* Chart */}
       <div css={{ height: 200 }}>
-        {traceCountDataPoints.length > 0 ? (
+        {hasData ? (
           <ResponsiveContainer width="100%" height="100%">
             <BarChart data={chartData} margin={{ top: 10, right: 20, left: 0, bottom: 0 }}>
               <XAxis dataKey="name" {...xAxisProps} />
